Refactor user stats into separate utility class.
